% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_addsite.R
\name{addsite_ui}
\alias{addsite_ui}
\title{UI Side for adding new trial site in HIDAP}
\usage{
addsite_ui(type = "tab", title = "New Trial Site", name = "createList")
}
\arguments{
\item{type}{type of UI element, deault is a tab in a shinydashboard}

\item{title}{diaply title name}

\item{name}{UI TabName}
}
\description{
Returns user friendly ui
}
\author{
Omar Benites
}
