% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_addsite_hdagrofims.R
\name{addsite_ui_agrofims}
\alias{addsite_ui_agrofims}
\title{UI Side for adding new trial site in for HIDAP-AGROFIMS}
\usage{
addsite_ui_agrofims(type = "tab", title = "New Trial Site",
  name = "createList")
}
\arguments{
\item{type}{type of UI element, default is a tab in a shinydashboard}

\item{title}{diaply title name}

\item{name}{UI TabName}
}
\description{
Returns user friendly ui for for HIDAP-AGROFIMS
}
\author{
Omar Benites
}
