% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_trial_tools.R
\name{filter_geodata}
\alias{filter_geodata}
\title{Get the geographical information from CIP trial sites}
\usage{
filter_geodata(sites_data, country_input, trial_site_abbr)
}
\arguments{
\item{sites_data}{The master trial sites dataset}

\item{country_input}{the input country-value using selectrizeInput
Function to get the whole list of countries from CIP trial tites}

\item{trial_site_abbr}{Trial site abbreviation}
}
\value{
vector
}
\description{
This function gives the geographical information
}
\author{
Omar Benites
}
