% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_trial_tools.R
\name{get_admin_agrofims}
\alias{get_admin_agrofims}
\title{Get the admin 1 for each country in AGROFIMS TRIALS}
\usage{
get_admin_agrofims(sites_data, country, admin1 = NULL, admin2 = NULL,
  admin3 = NULL, admin4 = NULL, admin5 = NULL)
}
\arguments{
\item{sites_data}{The master trial sites dataset}

\item{country}{the input country-value using selectrizeInput}

\item{admin1}{admin 1}

\item{admin2}{admin 2}

\item{admin3}{admin 3}

\item{admin4}{admin 4}

\item{admin5}{admin5
Function to get the whole list of countries from CIP trial tites}
}
\value{
vector
}
\description{
This function gives admins related to country
}
\author{
Omar Benites
}

