% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_trial_tools.R
\name{get_country_list}
\alias{get_country_list}
\title{Get the country list from CIP Master Trial Sites}
\usage{
get_country_list(sites_data)
}
\arguments{
\item{sites_data}{The master trial sites dataset}
}
\value{
vector
}
\description{
Function to get the list of countries of CIP trial sites
Function to get the whole list of countries from CIP trial tites
}
\author{
Omar Benites
}
