% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_trial_tools.R
\name{get_filter_locality}
\alias{get_filter_locality}
\title{Get the country list from CIP Master Trial Sites}
\usage{
get_filter_locality(sites_data, country_input)
}
\arguments{
\item{sites_data}{The master trial sites dataset}

\item{country_input}{the input country-value using selectrizeInput
Function to get the whole list of countries from CIP trial tites}
}
\value{
vector
}
\description{
This function gives the locality lists related to country
}
\author{
Omar Benites
}
