% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/api_fbsites.R
\name{get_site_table}
\alias{get_site_table}
\title{Gets a site table.}
\usage{
get_site_table()
}
\value{
dataframe
}
\description{
If not yet present creates a dummy one.
}
\author{
Reinhard Simon
}

