% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_fbsites.R
\name{post_site_table}
\alias{post_site_table}
\title{Posts a site table locally.}
\usage{
post_site_table(table_sites)
}
\arguments{
\item{table_sites}{data.frame}
}
\description{
Posts a data.frame containing location data.
}
\author{
Reinhard Simon
}

