% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{seg2dec}
\alias{seg2dec}
\title{Convertion to sexagesimal to decimal}
\usage{
seg2dec(grade = 0, minute = 0, second = 0)
}
\arguments{
\item{grade}{grades}

\item{minute}{minutes}

\item{second}{seconds}
}
\description{
Returns the convertion of sexagesimal system to decimal system.
}
\author{
Omar Benites, Franklin Plasencia
}
