% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_addsite.R
\name{server_addsite}
\alias{server_addsite}
\title{server_design}
\usage{
server_addsite(input, output, session, values)
}
\arguments{
\item{input}{shinyserver input}

\item{output}{shinyserver output}

\item{session}{shinyserver session}

\item{values}{reactive values}
}
\description{
Design a fieldbook
}
\author{
Omar Benites
}
