% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_addsite_hdagrofims.R
\name{server_addsite_agrofims}
\alias{server_addsite_agrofims}
\title{server_design for HIDAP-AGROFIMS}
\usage{
server_addsite_agrofims(input, output, session, values)
}
\arguments{
\item{input}{shinyserver input}

\item{output}{shinyserver output}

\item{session}{shinyserver session}

\item{values}{reactive values}
}
\description{
Design a fieldbook for HIDAP-AGROFIMS
}
\author{
Omar Benites
}

