% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/server_fbsites.R
\name{server_site}
\alias{server_site}
\title{Server side logic}
\usage{
server_site(input, output, session, dom = "hot_sites", values = values)
}
\arguments{
\item{input}{shinyserver input}

\item{output}{shinyserver output}

\item{session}{shinyserver session}

\item{dom}{target dom element name}

\item{values}{reactive values}
}
\description{
Constructs table
}
\author{
Reinhard Simon
}

