% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_fbsites.R
\name{ui_site}
\alias{ui_site}
\title{shiny UI element}
\usage{
ui_site(type = "tab", title = "List of trial sites",
  name = "resource_site", output = "hot_sites")
}
\arguments{
\item{type}{of ui Element; default is a tab in a shiny dashboard}

\item{title}{display title}

\item{name}{a reference name}

\item{output}{name of output element}
}
\description{
returns a re-usable user interface element
}
\author{
Reinhard Simon
}
