% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/iso_country.R
\docType{data}
\name{iso_country}
\alias{iso_country}
\title{ISO country data with continents and CIP regions}
\format{A data frame with 247 rows and 6 variables:
\describe{
  \item{country}{country name, English}
  \item{iso2}{ISO2 two letter abbreviation}
  \item{iso3}{ISO3 three letter abbreviation}
  \item{isoid}{ISO a sequential number}
  \item{continent}{continent}
  \item{cipregion}{to which CIP region the country belongs to}
}}
\source{
\url{http://www.nationsonline.org/oneworld/country_code_list.htm}
}
\usage{
iso_country
}
\description{
Contains world country names and classificatiosn
}
\keyword{datasets}

