% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ui_fbsites.R
\name{ui_site}
\alias{ui_site}
\title{shiny UI element}
\usage{
ui_site(type = "tab", title = "Site configuration",
  name = "resource_site")
}
\arguments{
\item{type}{of ui Element; default is a tab in a shiny dashboard}

\item{title}{display title}

\item{name}{a reference name}
}
\description{
returns a re-usable user interface element
}
\author{
Reinhard Simon
}

